Rename to poetry-plugin-pypi-proxy
